pyproject.toml
[project]
name = "datagenerator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "kafka-python>=2.0.2",
    "faker>=20.0.0",
    "requests>=2.31.0",
    "lz4>=4.4.5",
    "fastavro>=1.12.1",
]
